pyproject.toml
[tool.poetry]
name = "django-fitst"
version = "0.1.0"
description = ""
authors = ["azzip <azzipstar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
pydantic = "^2.4.2"
django-injector = "^0.3.1"
drf-yasg = "^1.21.7"
coreapi = "^2.3.3"
croniter = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
